financial-services/fraud-detection-agent/pyproject.toml
[project]
name = "fraud-detection-agent"
version = "0.1.0"
description = "LangGraph-based fraud detection and AML workflow sample"
requires-python = ">=3.11"
dependencies = [
    "langgraph==1.0.8",
    "langchain==1.2.10",
    "langchain-core==1.2.13",
    "langchain-anthropic==1.3.3",
    "langchain-community==0.4.1",
    "deepagents==0.4.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
